python/requirements.txt
# Python requirements for Data File Viewer extension
# Install with: pip install -r requirements.txt

numpy>=1.20.0
pandas>=1.3.0
h5py>=3.0.0
pyarrow>=5.0.0
msgpack>=1.0.0
joblib>=1.0.0
avro-python3>=1.10.0
netCDF4>=1.5.0
scipy>=1.7.0